cloudtests: rename files to avoid ImportPathMismatchError
