+++ misc/python/materialize/cloudtest/util/controller.py
@@ -15,7 +15,7 @@ from dataclasses import dataclass
 from typing import Any, List, Optional, Tuple, Union
 
 from materialize.cloudtest.util.common import eprint, retry
-from materialize.cloudtest.util.docker import docker_env
+from materialize.cloudtest.util.docker_env import docker_env
 
 
 @dataclass
